test: specs updated
