+++ Tests/pocketbase-swift-sdkTests/type_inference.swift
